control warning verbosity more closely
Suppresses warnings we don't care about by setting `rlib_warning_verbosity` to
"quiet" for the test suite. Then, sets to "verbose" or "default" for tests where
we are actually testing the warning behavior.

We had one test that would succeed on the first run but fail on subsequent runs
in the same R session because of not handling the verbosity properly.